index file was modified and store folder was created
